moved cohort creation code into its own class to make mpic_coordinator less bloated.
